remove favorites file
