+++ app/src/main/java/vanderclay/comet/benson/franticsearch/ui/adapters/FavoriteListAdapter.kt
@@ -34,40 +34,17 @@ class FavoriteListAdapter(val cards: MutableList<Card>): RecyclerView.Adapter<Ca
             val alertDialogBuilder = AlertDialog.Builder(holder.itemView?.context!!)
             alertDialogBuilder.setMessage("Delete Favorite? ${card.name}?")
             alertDialogBuilder.setPositiveButton("Yes", { _, _ ->
-                val favoriteCardReference = FirebaseDatabase
+                FirebaseDatabase
                         .getInstance()
                         .getReference("Favorites")
                         .child(FirebaseAuth.getInstance().currentUser?.uid)
                         .child(card.id)
                         .removeValue()
-                val cardRemoved = cards.removeAt(position)
+                cards.removeAt(position)
                 notifyItemRemoved(position)
                 notifyItemRangeChanged(position, cards.size)
                 notifyDataSetChanged()
 
-//                val valueEventListener = object : ValueEventListener {
-//                    override fun onDataChange(dataSnapshot: DataSnapshot) {
-//                        for (postSnapshot in dataSnapshot.children) {
-//                            FirebaseDatabase
-//                                    .getInstance()
-//                                    .getReference("Favorites")
-//                                    .child(FirebaseAuth.getInstance().currentUser?.uid)
-//                                    .child(postSnapshot.key)
-//                                    .removeValue()
-//                            val cardRemoved = cards.removeAt(position)
-//                            Log.d(TAG, cardRemoved.toString())
-//                            notifyItemRemoved(position)
-//                            notifyItemRangeChanged(position, cards.size)
-//                            notifyDataSetChanged()
-//                        }
-//
-//                    }
-//                    override fun onCancelled(databaseError: DatabaseError) {
-//                        Log.d(TAG, databaseError.toString())
-//                    }
-//                }
-//
-//                favoriteCardReference.addValueEventListener(valueEventListener)
             })
             alertDialogBuilder.setNegativeButton("No", { _, _ ->
                 Log.d(TAG, "Remove deck cancelled")
